Crate pallet_auto_id

Pallet AutoID

Re-exports

pub use weights::WeightInfo;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_auto_id

Structs

AutoId

A representation of AutoId

CertificateAction

Signing data used to verify the certificate action.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

RenewX509Certificate

Type holds X509 certificate details used to renew.

Signature

Signature holds algorithm used and the signature value.

X509Certificate

X509 certificate.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Certificate

Certificate associated with AutoId.

CertificateActionType

Specific action type taken by the subject of the Certificate.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

RegisterAutoId

Request to register a new AutoId.

RegisterAutoIdX509

Type holds X509 certificate details used to register an AutoId.

RenewAutoId

Request to renew AutoId.

Traits

Config

Configuration trait of this pallet.

Type Aliases

Identifier

Unique AutoId identifier.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Serial

Serial issued by the subject.